tests: Move editor test content to its own file
So it can be more easily shared between the posix and zx test
programs.
